stylesense-ai/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.30
aiofiles==23.2.1
python-dotenv==1.0.1
google-genai==0.5.0
pillow==10.3.0
httpx==0.27.0
jinja2==3.1.4
